dfs implementation trial
in this specific version we search  for mate in specific historic games
via looking for only moves that makes check, since it is either mate or
opponent can escape from us, if opponents move is sufficient for
escaping from subsequent checks and so mate we stop searching. This will
not be generelaizable and will work only for check..check..checkmate
types of puzzles.